+++ argilla-frontend/v1/domain/entities/field/Field.ts
@@ -16,15 +16,19 @@ export class Field {
     this.initializeOriginal();
   }
 
-  public get isTextType() {
-    return this.fieldType === "text";
+  get isTextType() {
+    return this.type === "text";
   }
 
-  private get fieldType() {
+  get isImageType() {
+    return this.type === "image";
+  }
+
+  get type() {
     return this.settings?.type?.toLowerCase() ?? null;
   }
 
-  public get isModified(): boolean {
+  get isModified(): boolean {
     return (
       this.title.trim() !== this.original.title ||
       this.settings.use_markdown !== this.original.settings.use_markdown
@@ -32,7 +36,7 @@ export class Field {
   }
 
   private MAX_TITLE_LENGTH = 500;
-  public validate(): Record<"title", string[]> {
+  validate(): Record<"title", string[]> {
     const validations: Record<"title", string[]> = {
       title: [],
     };
@@ -45,7 +49,7 @@ export class Field {
     return validations;
   }
 
-  public get isFieldValid(): boolean {
+  get isFieldValid(): boolean {
     return this.validate().title.length === 0;
   }