+++ argilla/src/argilla/_exceptions/__init__.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from argilla._exceptions._api import *  # noqa: F403
+from argilla._exceptions._client import *  # noqa: F403
 from argilla._exceptions._metadata import *  # noqa: F403
 from argilla._exceptions._serialization import *  # noqa: F403
 from argilla._exceptions._settings import *  # noqa: F403